dreamer/requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
Flask-Bcrypt